examples/custom_training_and_prediction_v2/serving/requirements.txt
joblib==1.4.0
numpy==1.26.4
scikit-learn==1.2.2
google-cloud-storage==2.16.0
